+++ src/EventListener/HookListener.php
@@ -10,10 +10,10 @@ namespace HeimrichHannot\EncoreBundle\EventListener;
 
 use Contao\CoreBundle\Framework\ContaoFramework;
 use Contao\LayoutModel;
-use Contao\Model;
 use Contao\PageModel;
 use Contao\PageRegular;
 use HeimrichHannot\EncoreBundle\Asset\EntrypointsJsonLookup;
+use HeimrichHannot\EncoreBundle\Asset\PageEntrypoints;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Twig\Environment;
 
@@ -30,14 +30,13 @@ class HookListener
     private $twig;
 
     /**
-     * @var EntrypointsJsonLookup
-     */
-    private $entrypointsJsonLookup;
-
-    /**
      * @var ContainerInterface
      */
     private $container;
+    /**
+     * @var PageEntrypoints
+     */
+    private $pageEntrypoints;
 
     /**
      * Constructor.
@@ -46,85 +45,57 @@ class HookListener
      * @param Environment           $twig
      * @param EntrypointsJsonLookup $entrypointsJsonLookup
      */
-    public function __construct(ContainerInterface $container, Environment $twig, EntrypointsJsonLookup $entrypointsJsonLookup)
+    public function __construct(ContainerInterface $container, Environment $twig, PageEntrypoints $pageEntrypoints)
     {
         $this->framework = $container->get('contao.framework');
         $this->twig = $twig;
-        $this->entrypointsJsonLookup = $entrypointsJsonLookup;
         $this->container = $container;
+        $this->pageEntrypoints = $pageEntrypoints;
     }
 
     /**
-     * Modify the page object.
+     * generatePage Hook
      *
-     * @param PageModel   $page
+     * @param PageModel $page
      * @param LayoutModel $layout
      * @param PageRegular $pageRegular
+     * @throws \Twig\Error\LoaderError
+     * @throws \Twig\Error\RuntimeError
+     * @throws \Twig\Error\SyntaxError
      */
-    public function addEncore(PageModel $page, LayoutModel $layout, PageRegular $pageRegular)
+    public function onGeneratePage(PageModel $page, LayoutModel $layout, PageRegular $pageRegular)
     {
-        $this->doAddEncore($page, $layout, $pageRegular);
+        $this->addEncore($page, $layout, $pageRegular);
+        $this->cleanGlobalArrays();
     }
 
-    public function doAddEncore(PageModel $page, LayoutModel $layout, PageRegular $pageRegular, ?string $encoreField = 'encoreEntries', bool $includeInline = false): void
+    /**
+     * @param PageModel $page
+     * @param LayoutModel $layout
+     * @param PageRegular $pageRegular
+     * @param string|null $encoreField
+     * @param bool $includeInline
+     * @throws \Twig\Error\LoaderError
+     * @throws \Twig\Error\RuntimeError
+     * @throws \Twig\Error\SyntaxError
+     */
+    public function addEncore(PageModel $page, LayoutModel $layout, PageRegular $pageRegular, ?string $encoreField = 'encoreEntries', bool $includeInline = false): void
     {
         if (!$layout->addEncore) {
             return;
         }
 
-        if (!$this->container->hasParameter('huh.encore')) {
-            return;
-        }
-
-        $config = $this->container->getParameter('huh.encore');
         $templateData = $layout->row();
 
-        // active entries
-        $jsEntries = [];
-        $cssEntries = [];
-        $jsHeadEntries = [];
-
-        // add entries from the entrypoints.json
-        if (isset($config['encore']['entrypointsJsons']) && \is_array($config['encore']['entrypointsJsons']) && !empty($config['encore']['entrypointsJsons'])) {
-            if (!isset($config['encore']['entries'])) {
-                $config['encore']['entries'] = [];
-            } elseif (!\is_array($config['encore']['entries'])) {
-                return;
-            }
-
-            $config['encore']['entries'] = $this->entrypointsJsonLookup->mergeEntries(
-                $config['encore']['entrypointsJsons'],
-                $config['encore']['entries'],
-                $layout->encoreBabelPolyfillEntryName,
-                $layout
-            );
-        }
-
-        if (!isset($config['encore']['entries']) || !\is_array($config['encore']['entries']) || empty($config['encore']['entries'])) {
+        if (!$this->pageEntrypoints->generatePageEntrypoints($page, $layout, $encoreField))
+        {
             return;
         }
 
-        foreach ($config['encore']['entries'] as $entry) {
-            if (!isset($entry['name'])) {
-                continue;
-            }
-            if ($this->isEntryActive($entry['name'], $layout, $page, $encoreField)) {
-                if (isset($entry['head']) && $entry['head']) {
-                    $jsHeadEntries[] = $entry['name'];
-                } else {
-                    $jsEntries[] = $entry['name'];
-                }
-
-                if (isset($entry['requiresCss']) && $entry['requiresCss']) {
-                    $cssEntries[] = $entry['name'];
-                }
-            }
-        }
-
-        $templateData['jsEntries'] = $jsEntries;
-        $templateData['jsHeadEntries'] = $jsHeadEntries;
-        $templateData['cssEntries'] = $cssEntries;
-
+        $templateData['jsEntries'] = $this->pageEntrypoints->getJsEntries();
+        $templateData['jsHeadEntries'] = $this->pageEntrypoints->getJsHeadEntries();
+        $templateData['cssEntries'] = $this->pageEntrypoints->getCssEntries();
+        
         // render css alone (should be used in <head>)
         $pageRegular->Template->encoreStylesheets = $this->twig->render(
             $this->getItemTemplateByName($layout->encoreStylesheetsImportsTemplate ?: 'default_css'), $templateData
@@ -159,56 +130,6 @@ class HookListener
         return false;
     }
 
-    /**
-     * @param string $entry
-     *
-     * @return bool
-     */
-    public function isEntryActive(string $entry, LayoutModel $layout, PageModel $currentPage, string $encoreField = 'encoreEntries'): bool
-    {
-        $parents = [$layout];
-
-        $parentPages = $this->container->get('huh.utils.model')->findParentsRecursively('pid', 'tl_page', $currentPage);
-        if (\is_array($parentPages)) {
-            $parents = array_merge($parents, $parentPages);
-        }
-
-        $result = false;
-
-        foreach (array_merge($parents, [$currentPage]) as $i => $page) {
-            $isActive = $this->isEntryActiveForPage($entry, $page, $encoreField);
-
-            if (0 == $i || null !== $isActive) {
-                $result = $isActive;
-            }
-        }
-
-        return $result ? true : false;
-    }
-
-    /**
-     * @param string $entry
-     * @param Model  $page
-     *
-     * @return bool|null Returns null, if no information about the entry is specified in the page; else bool
-     */
-    public function isEntryActiveForPage(string $entry, Model $page, string $encoreField = 'encoreEntries')
-    {
-        $entries = \Contao\StringUtil::deserialize($page->{$encoreField}, true);
-
-        foreach ($entries as $row) {
-            if ($row['entry'] === $entry) {
-                if ($page instanceof LayoutModel) {
-                    return true;
-                }
-
-                return $row['active'] ? true : false;
-            }
-        }
-
-        return null;
-    }
-
     public function cleanGlobalArrays()
     {
         if (!$this->container->get('huh.utils.container')->isFrontend()) {
@@ -225,14 +146,14 @@ class HookListener
             return;
         }
 
-        $config = $this->container->getParameter('huh.encore');
+        $config = $this->container->getParameter('huh_encore');
 
         // js
-        if (isset($config['encore']['legacy']['js']) && \is_array($config['encore']['legacy']['js'])) {
+        if (isset($config['legacy']['js']) && \is_array($config['legacy']['js'])) {
             $jsFiles = &$GLOBALS['TL_JAVASCRIPT'];
 
             if (\is_array($jsFiles)) {
-                foreach ($config['encore']['legacy']['js'] as $jsFile) {
+                foreach ($config['legacy']['js'] as $jsFile) {
                     if (isset($jsFiles[$jsFile])) {
                         unset($jsFiles[$jsFile]);
                     }
@@ -240,11 +161,11 @@ class HookListener
             }
         }
         // jquery
-        if (isset($config['encore']['legacy']['jquery']) && \is_array($config['encore']['legacy']['jquery'])) {
+        if (isset($config['legacy']['jquery']) && \is_array($config['legacy']['jquery'])) {
             $jqueryFiles = &$GLOBALS['TL_JQUERY'];
 
             if (\is_array($jqueryFiles)) {
-                foreach ($config['encore']['legacy']['jquery'] as $legacyFile) {
+                foreach ($config['legacy']['jquery'] as $legacyFile) {
                     if (isset($jqueryFiles[$legacyFile])) {
                         unset($jqueryFiles[$legacyFile]);
                     }
@@ -253,12 +174,12 @@ class HookListener
         }
 
         // css
-        if (isset($config['encore']['legacy']['css']) && \is_array($config['encore']['legacy']['css'])) {
+        if (isset($config['legacy']['css']) && \is_array($config['legacy']['css'])) {
             foreach (['TL_USER_CSS', 'TL_CSS'] as $arrayKey) {
                 $cssFiles = &$GLOBALS[$arrayKey];
 
                 if (\is_array($cssFiles)) {
-                    foreach ($config['encore']['legacy']['css'] as $cssFile) {
+                    foreach ($config['legacy']['css'] as $cssFile) {
                         if (isset($cssFiles[$cssFile])) {
                             unset($cssFiles[$cssFile]);
                         }
@@ -270,13 +191,13 @@ class HookListener
 
     public function getItemTemplateByName(string $name)
     {
-        $config = $this->container->getParameter('huh.encore');
+        $config = $this->container->getParameter('huh_encore');
 
-        if (!isset($config['encore']['templates']['imports'])) {
+        if (!isset($config['templates']['imports'])) {
             return null;
         }
 
-        $templates = $config['encore']['templates']['imports'];
+        $templates = $config['templates']['imports'];
 
         foreach ($templates as $template) {
             if ($template['name'] == $name) {