+++ src/Resources/contao/config/config.php
@@ -3,5 +3,4 @@
 /**
  * Hooks
  */
-$GLOBALS['TL_HOOKS']['generatePage']['huh.encore-bundle']       = ['huh.encore.listener.hooks', 'addEncore'];
-$GLOBALS['TL_HOOKS']['loadDataContainer']['huh.encore-bundle']  = ['huh.encore.listener.hooks', 'cleanGlobalArrays'];
+$GLOBALS['TL_HOOKS']['generatePage']['huh.encore-bundle']       = ['huh.encore.listener.hooks', 'onGeneratePage'];